requirements.txt
aggdraw==1.3.15
Pillow==10.1.0
